Include Iter8 list in the refactoring
